feat: adiciona validações de segurança nas funções

+++ services/deleteCategory.js
@@ -15,11 +15,25 @@ async function deleteCategory(user,rl,categoriesMenu,internalSystemMenu) {
 
     const [categories] = await connection.execute(sqlCategories);
 
+    if (categories.length === 0) {
+        console.log("Nenhuma categoria cadastrado! 🚫");
+        await pause(rl);
+        return categoriesMenu(user,rl,internalSystemMenu); 
+    }
+
     for (const category of categories) {
         console.log(`${category.id}. ${category.name}`);
     }
 
-    const selectCategory = await rl.question("\n📌 - Selecione o ID da categoria que deseja excluir: ");
+    const selectCategory = Number(await rl.question("\n📌 - Selecione o ID da categoria que deseja excluir: "));
+
+    const categoryExists = categories.find(category => category.id === selectCategory);
+
+    if (!categoryExists) {
+        console.log("\nCategoria não encontrada! 🚫"); 
+        await pause(rl);
+        return categoriesMenu(user,rl,internalSystemMenu);   
+    }
 
     const sqlDeleteCategory =
     `DELETE FROM categories
